Revert "Java2 to main"
