--- docker-compose.yml
@@ -1,34 +1,0 @@
-version: '3'
-
-volumes:
-  db-data:
-
-services:
-  # Postgres container / Production database
-  app_db:
-    image: postgres:latest
-
-    volumes:
-      - db-data:/var/lib/postgresql/data
-
-    environment:
-      - POSTGRES_DB=appointment_db
-      - POSTGRES_USER=postgres
-      - POSTGRES_PASSWORD=password
-      - PGDATA=/var/lib/postgresql/data/pgdata
-
-    ports:
-      - 5432:5432
-
-  # Web App Container
-  webapp:
-    build: .
-
-    environment:
-      - RUN_MODE=production
-
-    ports:
-      - 8080:8080
-
-    depends_on:
-      - app_db
