+++ pom.xml
@@ -3,18 +3,18 @@
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
-
-    <groupId>de.fhe.appointment</groupId>
-    <artifactId>Java2 Appointment</artifactId>
-    <packaging>jar</packaging>
+    <groupId>de.fhe.ai.appointment</groupId>
+    <artifactId>WS2021_Java_Team_1_Service_6_Appointments</artifactId>
+    <packaging>pom</packaging>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
         <java.version>11</java.version>
         <junit.version>5.4.0</junit.version>
+        <jsonassert.version>1.5.0</jsonassert.version>
         <jersey.version>3.0.2</jersey.version>
     </properties>
 
@@ -31,6 +31,7 @@
     </dependencyManagement>
 
     <dependencies>
+
         <!-- Jetty -->
         <dependency>
             <groupId>org.glassfish.jersey.containers</groupId>
@@ -64,19 +65,19 @@
             <version>2.0.1</version>
         </dependency>
 
-        <!-- Project Lombok -->
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-            <version>1.18.24</version>
-            <scope>provided</scope>
-        </dependency>
-
         <!-- JUnit 5 -->
         <dependency>
             <groupId>org.junit.jupiter</groupId>
             <artifactId>junit-jupiter-params</artifactId>
             <version>${junit.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <!-- test json data -->
+        <dependency>
+            <groupId>org.skyscreamer</groupId>
+            <artifactId>jsonassert</artifactId>
+            <version>${jsonassert.version}</version>
             <scope>test</scope>
         </dependency>
 
@@ -91,7 +92,7 @@
         <dependency>
             <groupId>org.hibernate</groupId>
             <artifactId>hibernate-entitymanager</artifactId>
-            <version>5.6.8.Final</version>
+            <version>5.6.9.Final</version>
         </dependency>
 
         <!-- Hibernate’s implementation of JSR-303. -->
@@ -114,10 +115,10 @@
             <version>1.7.36</version>
         </dependency>
         <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>4.13.2</version>
-            <scope>test</scope>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <version>RELEASE</version>
+            <scope>compile</scope>
         </dependency>
 
     </dependencies>
@@ -151,7 +152,7 @@
                         <manifest>
                             <addClasspath>true</addClasspath>
                             <classpathPrefix>lib/</classpathPrefix>
-                            <mainClass>de.fherfurt.appointment.WebApplication</mainClass>
+                            <mainClass>de.fhe.ai.addressbook.WebApplication</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
