+++ src/main/java/de/fhe/ai/appointment/model/SortCriteriaEnum.java
@@ -1,0 +1,13 @@
+package de.fhe.ai.appointment.model;
+
+public enum SortCriteriaEnum {
+    NAME_ASC,
+    NAME_DSC,
+    DATE_ASC,
+    DATE_DSC,
+    CREATOR_ASC,
+    CREATOR_DSC,
+    CAMPUSLOCATION_ASC,
+    CAMPUSLOCATION_DSC,
+    UNDEFINED
+}
