--- src/main/java/de/fherfurt/appointment/storage/repository/PersonDao.java
@@ -1,16 +1,0 @@
-package de.fherfurt.appointment.storage.repository;
-
-
-import de.fherfurt.appointment.models.Appointment;
-import de.fherfurt.appointment.models.Person;
-
-import java.util.Collection;
-
-/**
- * The interface Person dao.The interface Person dao which extends GenericDao<Person>
- *  presents the connection to Database.
- *  @author teichmueller
- */
-public interface PersonDao extends GenericDao<Person>{
-
-}
