--- src/test/java/de/fherfurt/appointment/Constants.java
@@ -1,33 +1,0 @@
-package de.fherfurt.appointment;
-
-public class Constants {
-    public static final String FIRST_PERSON_LASTNAME = "Mustermann";
-    public static final String FIRST_PERSON_FIRSTNAME = "Max";
-    public static final String FIRST_PERSON_MAIL = "m.m@muster.com";
-
-    public static final String SECOND_PERSON_LASTNAME = "Alm";
-    public static final String SECOND_PERSON_FIRSTNAME = "Heidi";
-    public static final String SECOND_PERSON_MAIL = "heidi@alm.ch";
-
-
-
-    public statsic final String FIRST_APPOINTMENT_NAME = "Java 2";
-    public statsic final String FIRST_APPOINTMENT_CREATOR = "Avemarg";
-    public statsic final Date FIRST_APPOINTMENT_DATE = "new Date()";                        //TODO: DATE FORMAT prüfen
-    public statsic final String FIRST_APPOINTMENT_REPITITION = "weekly";
-    public statsic final String FIRST_APPOINTMENT_STREET = "Altonaer Straße";
-    public statsic final String FIRST_APPOINTMENT_ROOM = "5.1.05";
-    public statsic final String FIRST_APPOINTMENT_DESCRIPTION = "Vorlesung Programmierung Java 2"
-                                                                                            // TODO: DATENTYP prüfen  --personList1));--
-
-    public statsic final String FIRST_APPOINTMENT_NAME = "Embedded Systems";
-    public statsic final String FIRST_APPOINTMENT_CREATOR = "Bischoff";
-    public statsic final Date FIRST_APPOINTMENT_DATE = "new Date()";                        //TODO: DATE FORMAT prüfen
-    public statsic final String FIRST_APPOINTMENT_REPITITION = "monthly";
-    public statsic final String FIRST_APPOINTMENT_STREET = "Altonaer Straße";
-    public statsic final String FIRST_APPOINTMENT_ROOM = "9.1.06";
-    public statsic final String FIRST_APPOINTMENT_DESCRIPTION = "Vorlesung Embedded Systems"
-                                                                                            // TODO: DATENTYP prüfen  --personList1));--
-
-
-}
